adapting new tests from the parent PR
