+++ reactor-core/src/test/java/reactor/core/publisher/OperatorsTest.java
@@ -17,6 +17,7 @@
 package reactor.core.publisher;
 
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 import java.util.concurrent.RejectedExecutionException;
 import java.util.concurrent.atomic.AtomicBoolean;
@@ -45,6 +46,7 @@ import reactor.test.util.RaceTestUtils;
 import reactor.util.context.Context;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatNullPointerException;
 
 public class OperatorsTest {
 
@@ -673,5 +675,31 @@ public class OperatorsTest {
 				.isNotNull()
 				.isSameAs(notScannable)
 				.isSameAs(rawFilterRef.get());
+	}
+
+	@Test
+	public void discardAdapterRejectsNull() {
+		assertThatNullPointerException().isThrownBy(() -> Operators.discardLocalAdapter(null, obj -> {}))
+		                                .as("type null check")
+		                                .withMessage("onDiscard must be based on a type");
+		assertThatNullPointerException().isThrownBy(() -> Operators.discardLocalAdapter(String.class, null))
+		                                .as("discardHook null check")
+		                                .withMessage("onDiscard must be provided a discardHook Consumer");
+	}
+
+	@Test
+	public void discardAdapterIsAdditive() {
+		List<String> discardOrder = Collections.synchronizedList(new ArrayList<>(2));
+
+		Function<Context, Context> first = Operators.discardLocalAdapter(Number.class, i -> discardOrder.add("FIRST"));
+		Function<Context, Context> second = Operators.discardLocalAdapter(Integer.class, i -> discardOrder.add("SECOND"));
+
+		Context ctx = first.apply(second.apply(Context.empty()));
+		Consumer<Object> test = ctx.getOrDefault(Hooks.KEY_ON_DISCARD, o -> {});
+
+		assertThat(test).isNotNull();
+
+		test.accept(1);
+		assertThat(discardOrder).as("consumers were combined").containsExactly("FIRST", "SECOND");
 	}
 }
